+++ skyportal/facility_apis/generic.py
@@ -7,7 +7,7 @@ from baselayer.app.env import load_env
 from baselayer.app.flow import Flow
 
 from ..utils import http
-from ..utils.UTCTZnaiveDateTime import utcnow_naive
+from ..utils.naive_datetime import utcnow_naive
 from . import FollowUpAPI
 
 env, cfg = load_env()
