+++ skyportal/facility_apis/sedm.py
@@ -10,7 +10,7 @@ from baselayer.app.flow import Flow
 from baselayer.log import make_log
 
 from ..utils import http
-from ..utils.UTCTZnaiveDateTime import utcnow_naive
+from ..utils.naive_datetime import utcnow_naive
 from . import FollowUpAPI, Listener
 
 env, cfg = load_env()
