+++ skyportal/facility_apis/swift.py
@@ -21,7 +21,7 @@ from baselayer.app.flow import Flow
 from baselayer.log import make_log
 
 from ..utils import http
-from ..utils.UTCTZnaiveDateTime import utcnow_naive
+from ..utils.naive_datetime import utcnow_naive
 from . import MMAAPI, FollowUpAPI
 
 env, cfg = load_env()
