+++ skyportal/facility_apis/tarot.py
@@ -16,7 +16,7 @@ from baselayer.log import make_log
 
 from ..utils import http
 from ..utils.calculations import get_next_valid_observing_time
-from ..utils.UTCTZnaiveDateTime import utcnow_naive
+from ..utils.naive_datetime import utcnow_naive
 from . import FollowUpAPI
 
 env, cfg = load_env()
