+++ skyportal/handlers/api/galaxy.py
@@ -31,7 +31,7 @@ from ...models import (
     Obj,
 )
 from ...utils.asynchronous import run_async
-from ...utils.UTCTZnaiveDateTime import utcnow_naive
+from ...utils.naive_datetime import utcnow_naive
 from ..base import BaseHandler, format_doc
 
 log = make_log("api/galaxy")
