+++ skyportal/handlers/api/sources_confirmed_in_gcn.py
@@ -13,7 +13,7 @@ from ...models import (
     Localization,
     SourcesConfirmedInGCN,
 )
-from ...utils.UTCTZnaiveDateTime import UTCTZnaiveDateTime
+from ...utils.naive_datetime import UTCTZnaiveDateTime
 from ..base import BaseHandler
 
 log = make_log("api/sources_confirmed_in_gcn")
