+++ skyportal/handlers/api/weather.py
@@ -6,8 +6,8 @@ from baselayer.app.access import auth_or_token
 from baselayer.app.env import load_env
 
 from ...models import Telescope, Weather
+from ...utils.naive_datetime import utcnow_naive
 from ...utils.offset import get_url
-from ...utils.UTCTZnaiveDateTime import utcnow_naive
 from ..base import BaseHandler
 
 _, cfg = load_env()
