+++ skyportal/models/facility_transaction.py
@@ -6,7 +6,7 @@ from sqlalchemy.orm import relationship
 
 from baselayer.app.models import Base
 
-from ..utils.UTCTZnaiveDateTime import utcnow_naive
+from ..utils.naive_datetime import utcnow_naive
 
 
 class FacilityTransaction(Base):
