+++ skyportal/models/phot_stat.py
@@ -22,7 +22,7 @@ from baselayer.app.models import (
 )
 from skyportal.models.photometry import PHOT_ZP, Photometry
 
-from ..utils.UTCTZnaiveDateTime import utcnow_naive
+from ..utils.naive_datetime import utcnow_naive
 
 _, cfg = load_env()
 PHOT_DETECTION_THRESHOLD = cfg["misc.photometry_detection_threshold_nsigma"]
