+++ skyportal/tests/api_tests/candidates_sources_events/test_sources.py
@@ -14,7 +14,7 @@ from tdtax import __version__, taxonomy
 from skyportal.models import cosmo
 from skyportal.tests import api
 
-from ....utils.UTCTZnaiveDateTime import utcnow_naive
+from ....utils.naive_datetime import utcnow_naive
 
 
 def test_source_list(view_only_token):
