+++ skyportal/tests/api_tests/test_reminders.py
@@ -4,7 +4,7 @@ from datetime import UTC, datetime, timedelta
 
 from skyportal.tests import api
 
-from ...utils.UTCTZnaiveDateTime import utcnow_naive
+from ...utils.naive_datetime import utcnow_naive
 
 
 def post_and_verify_reminder(endpoint, token):
