+++ skyportal/tests/frontend/test_reminders.py
@@ -6,7 +6,7 @@ from selenium.webdriver.common.by import By
 
 from skyportal.tests import api
 
-from ...utils.UTCTZnaiveDateTime import utcnow_naive
+from ...utils.naive_datetime import utcnow_naive
 
 
 def post_and_verify_reminder(endpoint, token):
